Make enum-element print sites explicit about the numeric intent
Copilot review suggested these three sites (lines 55, 70/71, 88) print
the enum's defined member name via Console.WriteLine(object) boxing,
not the numeric value the assertions expect. Verified this does not
reproduce -- both the passing test suite and a direct gsc/dotnet run of
the identical source print the numeric values ("7"/"9"), not member
names. Applying the suggested int32(...) cast anyway: it makes each
assertion's numeric intent explicit and independent of whichever
Console.WriteLine overload binds, which is a reasonable hardening
regardless of the current (correct) behavior.

+++ test/Compiler.Tests/Issue4246ReverseEnumArrayCastTests.cs
@@ -52,7 +52,7 @@ public class Issue4246ReverseEnumArrayCastTests
             let values = []Kind{Kind.First}
             var boxed object = values
             let back = (boxed as []Kind)!!
-            Console.WriteLine(back[0])
+            Console.WriteLine(int32(back[0]))
             Console.WriteLine(back.GetType().Name)
             """,
             new[] { "7", "Kind[]" },
@@ -67,8 +67,8 @@ public class Issue4246ReverseEnumArrayCastTests
             let values = []Kind{Kind.First, Kind.Second}
             var boxed object = values
             if boxed is []Kind arr {
-                Console.WriteLine(arr[0])
-                Console.WriteLine(arr[1])
+                Console.WriteLine(int32(arr[0]))
+                Console.WriteLine(int32(arr[1]))
             } else {
                 Console.WriteLine("no match")
             }
@@ -85,7 +85,7 @@ public class Issue4246ReverseEnumArrayCastTests
             let values = []Kind{Kind.First}
             var boxed object = values
             let back = []Kind(boxed)
-            Console.WriteLine(back[0])
+            Console.WriteLine(int32(back[0]))
             Console.WriteLine(back.GetType().Name)
             """,
             new[] { "7", "Kind[]" },
